Add Send via PD option to Slack "send message" actions
Conditionally stringify text field in Slack send-message-common.js

Refactor common elsements from Slack "send message" actions into
send-message-common.js
