+++ components/slack/actions/send-block-kit-message/send-block-kit-message.js
@@ -1,67 +1,33 @@
-const slack = require("../../slack.app.js");
+const common = require("../send-message-common");
 
 module.exports = {
+  ...common,
   key: "slack-send-block-kit-message",
   name: "Send Message Using Block Kit",
   description: "Send a message using Slack's Block Kit UI framework to a channel, group or user",
   version: "0.1.0",
   type: "action",
   props: {
-    slack,
+    ...common.props,
     conversation: {
       propDefinition: [
-        slack,
+        common.props.slack,
         "conversation",
       ],
       optional: false,
     },
     blocks: {
       propDefinition: [
-        slack,
+        common.props.slack,
         "blocks",
       ],
       optional: false,
     },
     text: {
       propDefinition: [
-        slack,
+        common.props.slack,
         "notificationText",
       ],
     },
-    as_user: {
-      propDefinition: [
-        slack,
-        "as_user",
-      ],
-    },
-    username: {
-      propDefinition: [
-        slack,
-        "username",
-      ],
-    },
-    icon_emoji: {
-      propDefinition: [
-        slack,
-        "icon_emoji",
-      ],
-    },
-    icon_url: {
-      propDefinition: [
-        slack,
-        "icon_url",
-      ],
-    },
-  },
-  async run() {
-    return await this.slack.sdk().chat.postMessage({
-      text: this.text,
-      blocks: this.blocks,
-      channel: this.conversation,
-      as_user: this.as_user,
-      username: this.username,
-      icon_emoji: this.icon_emoji,
-      icon_url: this.icon_url,
-    });
   },
 };
