+++ components/slack/actions/send-custom-message/send-custom-message.js
@@ -1,121 +1,79 @@
-const slack = require("../../slack.app.js");
+const common = require("../send-message-common");
 
 module.exports = {
+  ...common,
   key: "slack-send-custom-message",
   name: "Send a Custom Message",
   description: "Customize advanced setttings and send a message to a channel, group or user",
   version: "0.1.0",
   type: "action",
   props: {
-    slack,
+    ...common.props,
     conversation: {
       propDefinition: [
-        slack,
+        common.props.slack,
         "conversation",
       ],
     },
     text: {
       propDefinition: [
-        slack,
+        common.props.slack,
         "text",
       ],
     },
     attachments: {
       propDefinition: [
-        slack,
+        common.props.slack,
         "attachments",
       ],
     },
     unfurl_links: {
       propDefinition: [
-        slack,
+        common.props.slack,
         "unfurl_links",
       ],
     },
     unfurl_media: {
       propDefinition: [
-        slack,
+        common.props.slack,
         "unfurl_media",
       ],
     },
     parse: {
       propDefinition: [
-        slack,
+        common.props.slack,
         "parse",
-      ],
-    },
-    as_user: {
-      propDefinition: [
-        slack,
-        "as_user",
-      ],
-    },
-    username: {
-      propDefinition: [
-        slack,
-        "username",
-      ],
-    },
-    icon_emoji: {
-      propDefinition: [
-        slack,
-        "icon_emoji",
-      ],
-    },
-    icon_url: {
-      propDefinition: [
-        slack,
-        "icon_url",
       ],
     },
     mrkdwn: {
       propDefinition: [
-        slack,
+        common.props.slack,
         "mrkdwn",
       ],
     },
     blocks: {
       propDefinition: [
-        slack,
+        common.props.slack,
         "blocks",
       ],
     },
     link_names: {
       propDefinition: [
-        slack,
+        common.props.slack,
         "link_names",
       ],
     },
     reply_broadcast: {
       propDefinition: [
-        slack,
+        common.props.slack,
         "reply_broadcast",
       ],
     },
     thread_ts: {
       propDefinition: [
-        slack,
+        common.props.slack,
         "thread_ts",
       ],
     },
-  },
-  async run() {
-    return await this.slack.sdk().chat.postMessage({
-      text: this.text,
-      channel: this.conversation,
-      attachments: this.attachments,
-      unfurl_links: this.unfurl_links,
-      unfurl_media: this.unfurl_media,
-      parse: this.parse,
-      as_user: this.as_user,
-      username: this.username,
-      icon_emoji: this.icon_emoji,
-      icon_url: this.icon_url,
-      mrkdwn: this.mrkdwn,
-      blocks: this.blocks,
-      link_names: this.link_names,
-      reply_broadcast: this.reply_broadcast,
-      thread_ts: this.thread_ts,
-    });
   },
 };
