+++ components/slack/actions/send-message-private-channel/send-message-private-channel.js
@@ -1,4 +1,4 @@
-const slack = require("../../slack.app.js");
+const common = require("../send-message-common");
 
 module.exports = {
   key: "slack-send-message-private-channel",
@@ -7,55 +7,39 @@ module.exports = {
   version: "0.1.0",
   type: "action",
   props: {
-    slack,
+    ...common.props,
     conversation: {
       propDefinition: [
-        slack,
+        common.props.slack,
         "privateChannel",
       ],
     },
     text: {
       propDefinition: [
-        slack,
+        common.props.slack,
         "text",
-      ],
-    },
-    as_user: {
-      propDefinition: [
-        slack,
-        "as_user",
       ],
     },
     username: {
       propDefinition: [
-        slack,
+        common.props.slack,
         "username",
       ],
       description: "Optionally customize your bot's username (default is `Pipedream`).",
     },
     icon_emoji: {
       propDefinition: [
-        slack,
+        common.props.slack,
         "icon_emoji",
       ],
       description: "Optionally use an emoji as the bot icon for this message (e.g., `:fire:`). This value overrides `icon_url` if both are provided.",
     },
     icon_url: {
       propDefinition: [
-        slack,
+        common.props.slack,
         "icon_url",
       ],
       description: "Optionally provide an image URL to use as the bot icon for this message.",
     },
-  },
-  async run() {
-    return await this.slack.sdk().chat.postMessage({
-      channel: this.conversation,
-      text: this.text,
-      as_user: this.as_user,
-      username: this.username,
-      icon_emoji: this.icon_emoji,
-      icon_url: this.icon_url,
-    });
   },
 };
